+++ src/components/logo.tsx
@@ -1,0 +1,22 @@
+import { useMemo } from "react";
+import { useTheme } from "./theme-provider";
+
+interface Props{
+     width: number,
+     height: number,
+}
+export default function Logo({width, height}: Props){
+     const {theme} = useTheme()
+     const logoImage = useMemo(() => {
+          return theme === "light" ? "/logos/logo-green.webp" : "/logos/logo-lime.webp";
+     }, [theme]);
+     return (
+          <img
+               src={logoImage}
+               alt="Calm Mood"
+               width={width}
+               height={height}
+               className="object-contain"
+          />
+     )
+}