+++ src/components/theme-provider.tsx
@@ -1,0 +1,98 @@
+import { createContext, useContext, useEffect, useMemo, useState } from "react"
+
+type Theme = "dark" | "light" | "system"
+
+type ThemeProviderProps = {
+  children: React.ReactNode
+  defaultTheme?: Theme
+  storageKey?: string
+}
+
+type ThemeProviderState = {
+  theme: Theme
+  resolvedTheme: "dark" | "light" | null,
+  setTheme: (theme: Theme) => void
+}
+
+const initialState: ThemeProviderState = {
+  theme: "system",
+  resolvedTheme: null,
+  setTheme: () => null,
+}
+
+const ThemeProviderContext = createContext<ThemeProviderState>(initialState)
+
+export function ThemeProvider({
+  children,
+  defaultTheme = "system",
+  storageKey = "vite-ui-theme",
+  ...props
+}: ThemeProviderProps) {
+  const [theme, setTheme] = useState<Theme>(
+    () => (localStorage.getItem(storageKey) as Theme) || defaultTheme
+  )
+  const [resolvedTheme, setResolvedTheme] = useState<"dark" | "light">(
+    () => {
+      if (typeof window === "undefined") return "light"
+      return window.matchMedia("(prefers-color-scheme: dark)").matches
+        ? "dark"
+        : "light"
+    }
+  )
+
+  useEffect(() => {
+    const root = window.document.documentElement
+    root.classList.remove("light", "dark")
+
+    if (theme === "system") {
+      const systemTheme = window.matchMedia("(prefers-color-scheme: dark)").matches
+        ? "dark"
+        : "light"
+
+      root.classList.add(systemTheme)
+      setResolvedTheme(systemTheme)
+    } else {
+      root.classList.add(theme)
+      setResolvedTheme(theme)
+    }
+  }, [theme])
+  useEffect(() => {
+    if (theme !== "system") return
+
+    const media = window.matchMedia("(prefers-color-scheme: dark)")
+
+    const handler = () => {
+      const newTheme = media.matches ? "dark" : "light"
+      document.documentElement.classList.remove("light", "dark")
+      document.documentElement.classList.add(newTheme)
+      setResolvedTheme(newTheme)
+    }
+
+    media.addEventListener("change", handler)
+    return () => media.removeEventListener("change", handler)
+  }, [theme])
+
+  const value = useMemo(()=>({
+    theme,
+    resolvedTheme,
+    setTheme: (theme: Theme) => {
+      localStorage.setItem(storageKey, theme)
+      setTheme(theme)
+    },
+  }),[theme, resolvedTheme])
+
+  return (
+    <ThemeProviderContext.Provider {...props} value={value}>
+      {children}
+    </ThemeProviderContext.Provider>
+  )
+}
+
+export const useTheme = () => {
+  const context = useContext(ThemeProviderContext)
+
+  if (context === undefined)
+    throw new Error("useTheme must be used within a ThemeProvider")
+
+  return context
+}