+++ src/main.tsx
@@ -1,9 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App";
+import { ThemeProvider } from "./components/theme-provider";
+import { RouterProvider } from "react-router";
+import { router } from "./routes";
+import "./App.css";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
-    <App />
+    <ThemeProvider storageKey="calm-mood-theme">
+      <RouterProvider router={router}/>
+    </ThemeProvider>
   </React.StrictMode>,
 );