+++ src/pages/exercise.tsx
@@ -1,0 +1,9 @@
+import WindowWrapper from "@/components/window";
+
+export default function BreathingExercisePage(){
+     return (
+          <WindowWrapper title="Շնչառական վարժություն">
+               Hello New Calm World, Goodbye Hellish world!
+          </WindowWrapper>
+     )
+}