+++ src/pages/pomodoro.tsx
@@ -1,0 +1,9 @@
+import WindowWrapper from "@/components/window";
+
+export default function PomodoroPage(){
+     return (
+          <WindowWrapper title="Պոմոդորոյի մեթոդ">
+               Hello New Calm World, Goodbye Hellish world!
+          </WindowWrapper>
+     )
+}