+++ src/pages/sounds.tsx
@@ -1,0 +1,9 @@
+import WindowWrapper from "@/components/window";
+
+export default function SoundsPage(){
+     return (
+          <WindowWrapper title="Հանգստացնող ձայններ">
+               Hello New Calm World, Goodbye Hellish world!
+          </WindowWrapper>
+     )
+}