+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes of Calm Mood Desktop will be documented here.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - Unreleased
+### Added
+- Theme Changer
+- Splash Screen
+### Improved
+- Window Wrapper - Added more buttons with similar design
+
 ## [0.1.0] - 2026-06-10
 > [!IMPORTANT]
 > The first pre-release is mostly focused on the wrapper design meaning that it won't work or some parts might work